pyproject.toml
[project]
name = "nonebot-plugin-zhihu-parser"
version = "0.1.0"
description = "Nonebot2 知乎解析插件 (从 AstrBot 移植)"
authors = [
  {name = "devil", email = "devil@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-adapter-onebot>=2.4.0",
    "curl_cffi>=0.5.10",
    "beautifulsoup4>=4.12.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.nonebot]
plugins = ["nonebot_plugin_zhihu_parser"]